Struct libra_management::storage::StorageWrapper

pub struct StorageWrapper { /* fields omitted */ }

A helper to handle common error handling and functionality for tooling

Implementations

impl StorageWrapper

pub fn new(
storage_name: &'static str,
backend: &SecureBackend
) -> Result<StorageWrapper, Error>

pub fn new_with_namespace(
storage_name: &'static str,
namespace: String,
backend: &SecureBackend
) -> Result<StorageWrapper, Error>

pub fn value(&self, name: &'static str) -> Result<Value, Error>

pub fn account_address(
&self,
name: &'static str
) -> Result<AccountAddress, Error>

pub fn string(&self, name: &'static str) -> Result<String, Error>

pub fn transaction(&self, name: &'static str) -> Result<Transaction, Error>

pub fn u64(&self, name: &'static str) -> Result<u64, Error>

pub fn waypoint(&self, name: &'static str) -> Result<Waypoint, Error>

pub fn ed25519_key(&self, name: &'static str) -> Result<Ed25519PublicKey, Error>

Retrieves the Public key, that is stored as a public key

pub fn x25519_key(&self, name: &'static str) -> Result<PublicKey, Error>

Retrieves the Public key that is stored as a public key

pub fn rotate_key(
&mut self,
name: &'static str
) -> Result<Ed25519PublicKey, Error>

pub fn ed25519_public_from_private(
&self,
key_name: &'static str
) -> Result<Ed25519PublicKey, Error>

Retrieves public key from the stored private key

pub fn x25519_public_from_private(
&self,
key_name: &'static str
) -> Result<PublicKey, Error>

Retrieves public key from the stored private key

pub fn set(&mut self, name: &'static str, value: Value) -> Result<(), Error>

pub fn sign(
&mut self,
key_name: &'static str,
script_name: &'static str,
raw_transaction: RawTransaction
) -> Result<SignedTransaction, Error>

Sign a transaction