Module libra_management::validator_config

Functions

create_transaction

create_validator_config_and_reconfigure_script

create_validator_config_script

decode_validator_address

encode_address
Encode an address into bytes

encode_validator_address
Encodes and encrypts a validator address

fetch_keys_from_storage

owner_from_key
Retrieves the owner key from the remote storage using the owner name given by the validator-config command, and uses this key to derive an owner account address. If a remote storage path is not specified, returns an error.

update_config_and_reconfigure